+++ src/examples/index.ts
@@ -1,2 +1,2 @@
+export { default as TestBatch } from './TestBatch';
 export { default as TestBase } from './TestBase';
-export { default as TestPreload } from './TestPreload';
